tooltip on in-place functions, julia lang hint in metadata
ship (for now) our own version of tooltip.js to have tooltip on mutating
function:

    permute!(<tab> now show a tooltip in notebook.

add a

    "language" : "Julia"

Hint in metadata to help distinguishing .ipynb file that are python that
the julia one for external program. (nbviewer/nbconvert could be modify
to use that and have correct syntax highlighting later). This would be
deprecated in favor of an official way to set the language in later
notebooks versions.