Delay sending the idle message.
Jupyter removes IO handling callbacks associated with a cell when the `"status":"idle"` message with the (parent header's) `msg_id` associated with the cell is received. New (asynchronous) output stream messages received after the `"status":"idle"` is received will still be routed to the previously active cell under normal circumstances. However, if a new message is sent by the front-end, the `last_msg_callbacks` that would have put the output on an appropriate cell are overwritten/cleared and subsequent stream messages are silently dropped.

This change means IJulia will delay sending the `status:idle`. Among potentially other cases, it will help when Interact.jl displays a signal but also has an error during an asynchronous `push!` to that or another signal. In that case Interact's javascript sends a comm_msg which clears the  `last_msg_callbacks` as described above, and the error message/stacktrace is usually truncated. With this change, IJulia will delay sending the idle by and the stderr messages will be displayed.