Refactor heartbeat to shutdown cleanly (#1135)
* Refactor heartbeat to shutdown cleanly

From ZMQ docs: "zmq_proxy() runs in the current thread and returns only
if/when the current context is closed."

The heartbeat socket doesn't need to be global, as nothing else touches
it. BUT, if we create the heartbeat socket in a `Context` that has a global ref,
we can close the context, which will cause zmq_proxy to return and then
that thread to end/finish.

Doing that before shutting down helps avoid a segfault on shutdown.

* Update src/heartbeat.jl

Co-authored-by: Steven G. Johnson <stevenj@mit.edu>

* Create heartbeat and context in `init.jl`

---------

Co-authored-by: Steven G. Johnson <stevenj@mit.edu>