Attempt to send stdio to the kernel before sending execute_reply
The Jupyter messaging docs state that all messages on the shell (request)
channel shall be ended with a `status: idle` message, and that: "This idle
status message indicates that IOPub messages associated with a given request
have all been received."

Previously we did not attempt to ensure this, which meant that in some cases the
`execute_result` message would be published after the `execute_reply` message
had been sent. It's still not exactly guaranteed by the `yield()`, but it's
something and it makes the `jupyter_kernel_test` tests pass.