Allow registering hooks at any time
Technically it was also not allowed beforehand but in practice some packages do
so (e.g. RCall.jl in tests) and there doesn't seem to be a reason to forbid
it. Also put the functions in the public API section of the docs, which seems
fine since they're also documented in the module docstring.