change from using a build script to auto installing kernels and providing a jupyter path update function
The Pkg build functionality is not good. It gets run whenever a package is *downloaded* (which is a weird thing to trigger on where packages are considered immutable and multiple environments share the same package downloads), it requires the package to recompile when build options are changed etc.

This instead auto installs the kernel if needed and also provides a separate function to update the jupyter path.

It also moves to using the Preference system over using `DEPOT_PATH[1]/prefs` (but reads the old file as a fallback)

It also changes some things that had to be done at build time (like setting the debug mode) to now just be a runtime thing